+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Home screen day title flow now converts all currencies into primary
 * Markdown editor (transaction description) now has a preview, and minimal toolbar
+* Now you can duplicate a transaction be swiping to the right, closes [#232](https://github.com/flow-mn/flow/issues/232)
 
 ### Changes
 
@@ -25,6 +26,8 @@
 * Transaction list tile title color is now fixed in light themes
 * Wavy divider color now follows the theme change
 * Disabled Autocorrect on transaction title, so it no longer alters suggestions
+* Category and account detail page now doesn't include pending transactions in
+  count and flow.
 
 ## Beta 0.8.1